% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B3_data.R
\docType{data}
\name{B3_data}
\alias{B3_data}
\title{B3 Artificial Dataset}
\format{
A data frame with 2300 data points and 3 variables
\describe{
  \item{\code{x}}{Numeric values generated from Gaussian distributions}
  \item{\code{y}}{Numeric values generated from Gaussian distributions}
  \item{\code{label}}{Categorical labels 1,2,3,4,5}
}
}
\usage{
B3_data
}
\description{
A \code{2}-dimensional dataset from Wiroonsri and Preedasawakul (2024) generated from \code{5} different Gaussian distributions labeled as \code{1-5}.
}
\references{
{O. Preedasawakul, and N. Wiroonsri, A Bayesian Cluster Validity Index, Computational Statistics & Data Analysis, 202, 108053, 2025. \doi{10.1016/j.csda.2024.108053}}
}
\seealso{
{\link{B2_data}, \link{B4_data}, \link{B_WP.IDX},  \link{B_Wvalid}, \link{B_XB.IDX}}
}
\author{
{Nathakhun Wiroonsri and Onthada Preedasawakul}
}
\keyword{datasets}
