% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B1_data.R
\docType{data}
\name{B1_data}
\alias{B1_data}
\title{B1 Artificial Dataset}
\format{
A data frame with 5500 data points and 3 variables
\describe{
  \item{\code{x}}{Numeric values generated from Gaussian and Uniform distributions}
  \item{\code{y}}{Numeric values generated from Gaussian and Uniform distributions}
  \item{\code{label}}{Categorical labels 1,2}
}
}
\usage{
B1_data
}
\description{
A \code{2}-dimensional dataset from Wiroonsri and Preedasawakul (2024) generated from \code{1} Gaussian and \code{1} Uniform distributions labeled as \code{1-2}.
}
\references{
{N. Wiroonsri, O. Preedasawakul, A Bayesian cluster validity index, arXiv:2402.02162, 2024}
}
\seealso{
{\link{B2_data}, \link{B3_data}, \link{B_WP.IDX},  \link{B_Wvalid}, \link{B_XB.IDX}}
}
\author{
{Nathakhun Wiroonsri and Onthada Preedasawakul}
}
\keyword{datasets}
