% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B2_data.R
\docType{data}
\name{B2_data}
\alias{B2_data}
\title{B2 Artificial Dataset}
\format{
A data frame with 850 data points and 3 variables
\describe{
  \item{\code{x}}{Numeric values generated from Gaussian distributions}
  \item{\code{y}}{Numeric values generated from Gaussian distributions}
  \item{\code{label}}{Categorical labels 1,2,3,4,5}
}
}
\usage{
B2_data
}
\description{
A \code{2}-dimensional dataset from Wiroonsri and Preedasawakul (2024) generated from \code{5} different Gaussian distributions labeled as \code{1-5}.
}
\references{
{N. Wiroonsri, O. Preedasawakul, A Bayesian cluster validity index, arXiv:2402.02162, 2024}
}
\seealso{
{\link{B1_data}, \link{B3_data}, \link{B_WP.IDX},  \link{B_Wvalid}, \link{B_XB.IDX}}
}
\author{
{Nathakhun Wiroonsri and Onthada Preedasawakul}
}
\keyword{datasets}
