% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B5_data.R
\docType{data}
\name{B5_data}
\alias{B5_data}
\title{B5 Artificial Dataset}
\format{
A data frame with 1820 data points and 3 variables
\describe{
  \item{\code{x}}{Numeric values generated from Gaussian and Uniform distributions}
  \item{\code{y}}{Numeric values generated from Gaussian and Uniform distributions}
  \item{\code{label}}{Categorical labels 1,2,3,4,5,6,7,8,9}
}
}
\usage{
B5_data
}
\description{
A \code{2}-dimensional dataset from Wiroonsri and Preedasawakul (2024) generated from \code{7} different Gaussian and \code{2} Uniform distributions labeled as \code{1-9}.
}
\references{
{N. Wiroonsri, O. Preedasawakul, A Bayesian cluster validity index, arXiv:2402.02162, 2024}
}
\seealso{
{\link{B4_data}, \link{B6_data}, \link{B_WP.IDX},  \link{B_Wvalid}, \link{B_XB.IDX}}
}
\author{
{Nathakhun Wiroonsri and Onthada Preedasawakul}
}
\keyword{datasets}
